+++ test/helpers/index.js
@@ -2,7 +2,9 @@ const tape = require('tape')
 const HyperDHT = require('@hyperswarm/dht')
 const Hyperswarm = require('../../')
 
-module.exports = { test, swarm, destroy, defer }
+const CONNECTION_TIMEOUT = 100
+
+module.exports = { test, swarm, destroy, destroyAll, timeoutPromise, planPromise, defer }
 test.only = (name, fn) => test(name, fn, true, false)
 test.skip = (name, fn) => test(name, fn, false, true)
 
@@ -62,5 +64,65 @@ async function test (name, fn, only = false, skip = false) {
 
     destroy(bootstrappers)
     destroy(nodes)
+  }
+}
+
+async function destroyAll (...swarms) {
+  for (const swarm of swarms) {
+    await swarm.clear()
+  }
+  for (const swarm of swarms) {
+    await swarm.destroy()
+  }
+}
+
+function timeoutPromise (ms = CONNECTION_TIMEOUT) {
+  let res = null
+  let rej = null
+  let timer = null
+
+  const p = new Promise((resolve, reject) => {
+    res = resolve
+    rej = reject
+  })
+  p.resolve = res
+  p.reject = rej
+  p.reset = () => {
+    if (timer) clearTimeout(timer)
+    timer = setTimeout(() => p.reject(new Error('Timed out')), ms)
+  }
+
+  p.reset()
+  return p
+}
+
+function planPromise (t, count) {
+  let tick = 0
+  let res = null
+  let rej = null
+  let timer = null
+
+  let p = new Promise((resolve, reject) => {
+    res = resolve
+    rej = reject
+  })
+
+  return {
+    plan: (n) => {
+      const promise = planPromise(t, n)
+      p = Promise.all([p, promise])
+      return promise
+    },
+    pass: (msg) => {
+      t.pass(msg)
+      if (++tick === count) {
+        if (timer) clearTimeout(timer)
+        return res()
+      }
+    },
+    timeout: (ms) => {
+      timer = setTimeout(rej, ms, new Error('Plan promise timed out'))
+    },
+    then: (...args) => p.then(...args)
   }
 }
